# perf(codec): improve bytify and hexify perf