--- a/.changeset/big-ears-allow.md
+++ b/.changeset/big-ears-allow.md
@@ -1,0 +1,5 @@
+---
+"@ckb-lumos/codec": patch
+---
+
+improving `hexify` and `bytify` performance