# Description

A performance issue was found in #579 

## Type of change

- [x] Refactor (non-breaking change)

# How Has This Been Tested?

<details>
<summary>benchmark.js</summary>

```js
import Benchmark from "benchmark";
import { randomBytes } from "node:crypto";
import bytesOld from "./lib/bytes.js";
import bytesNew from "./lib/bytes-new.js";

const bytesData = randomBytes(10000);
const hexData = bytesNew.hexify(bytesData);

const suite = Benchmark.Suite();

suite
  .add("hexify-old", () => {
    bytesOld.hexify(bytesData);
  })
  .add("hexify-new", () => {
    bytesNew.hexify(bytesData);
  })
  .add("bytify-old", () => {
    bytesOld.bytify(hexData);
  })
  .add("bytify-new", () => {
    bytesNew.bytify(hexData);
  })
  .on("cycle", (event) => {
    console.log(String(event.target));
  })
  .run();
```

</details>

```
➜  codec git:(perf-bytes) ✗ bun run bench.mjs
hexify-old x 1,328 ops/sec ±0.38% (95 runs sampled)
hexify-new x 4,606 ops/sec ±0.33% (95 runs sampled)
bytify-old x 1,192 ops/sec ±0.51% (95 runs sampled)
bytify-new x 4,597 ops/sec ±0.29% (96 runs sampled)
➜  codec git:(perf-bytes) ✗ clear
➜  codec git:(perf-bytes) ✗ bun run bench.mjs
hexify-old x 1,325 ops/sec ±0.36% (95 runs sampled)
hexify-new x 4,524 ops/sec ±0.30% (97 runs sampled)
bytify-old x 1,184 ops/sec ±0.42% (96 runs sampled)
bytify-new x 4,547 ops/sec ±0.31% (95 runs sampled)
➜  codec git:(perf-bytes) ✗ node bench.mjs
hexify-old x 2,079 ops/sec ±0.75% (96 runs sampled)
hexify-new x 21,727 ops/sec ±1.01% (93 runs sampled)
bytify-old x 2,320 ops/sec ±0.29% (97 runs sampled)
bytify-new x 2,440 ops/sec ±0.38% (98 runs sampled)
➜  codec git:(perf-bytes) ✗ node bench.mjs
hexify-old x 2,092 ops/sec ±0.74% (93 runs sampled)
hexify-new x 22,277 ops/sec ±0.49% (99 runs sampled)
bytify-old x 2,319 ops/sec ±0.39% (97 runs sampled)
bytify-new x 2,467 ops/sec ±0.17% (99 runs sampled)
➜  codec git:(perf-bytes) ✗ node --version
v21.3.0
➜  codec git:(perf-bytes) ✗ bun --version    
1.0.15
```

<!-- 
# Checklist:

- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [x] I have added tests that prove my fix is effective or that my feature works
- [x] New and existing unit tests pass locally with my changes
- [x] Any dependent changes have been merged and published in downstream modules -->